# aw05-Price1999a/pos-delivery/src/main/java/com/micropos/delivery/model/Item.java
package com.micropos.delivery.model;

import lombok.AllArgsConstructor;
import lombok.Data;

import java.io.Serializable;

@Data
@AllArgsConstructor
public class Item implements Serializable {
    private Product product;
    private int quantity;
}