# aw05-Price1999a/pos-delivery/src/main/java/com/micropos/delivery/model/Product.java
package com.micropos.delivery.model;

import lombok.AllArgsConstructor;
import lombok.Data;

import java.io.Serializable;

@Data
@AllArgsConstructor
public class Product implements Serializable {
    private String id;
    private String name;
    private double price;
    private String image;
}