# aw05-Price1999a/pos-gateway/src/main/java/com/micropos/gateway/model/Item.java
package com.micropos.gateway.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.io.Serializable;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Item implements Serializable {
    private Product product;
    private int quantity;
}