# aw05-Price1999a/pos-gateway/src/main/java/com/micropos/gateway/model/Product.java
package com.micropos.gateway.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.io.Serializable;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Product implements Serializable {
    private String id;
    private String name;
    private double price;
    private String image;
}